+++ Changes.md
@@ -5,6 +5,8 @@
 
 * lint: add 'prefer-template'
 * restore TLS version info, set correctly #2723
+* add JSON format for logging
+
 
 ### New features